Extracted the tutorialVC into it's own class
